+++ package.json
@@ -17,6 +17,7 @@
   },
   "homepage": "https://github.com/retroplasma/earth-reverse-engineering#readme",
   "dependencies": {
-    "bmp-js": "^0.1.0"
+    "bmp-js": "^0.1.0",
+    "decode-dxt": "^1.0.1"
   }
 }
